meeting/requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.8
python-dateutil>=2.9
dateparser>=1.2
pytz>=2024.1
motor>=3.4
pymongo>=4.8
google-api-python-client>=2.146
google-auth>=2.35
google-auth-oauthlib>=1.2
langchain>=0.3
google-generativeai>=0.7
mcp>=1.1
httpx>=0.27
python-dotenv>=1.0
